refactor: move custom view components from app-shell to app-shell-connectors package
