+++ packages/application-components/src/components/custom-views/custom-views-selector/fetch-custom-views-by-locator.settings.graphql
