+++ packages/application-shell/src/components/custom-view-dev-host/custom-view-dev-host.tsx
@@ -1,5 +1,6 @@
 import { type ReactNode, useState } from 'react';
 import { Route, Router, Switch, useRouteMatch } from 'react-router-dom';
+import { CustomViewLoader } from '@commercetools-frontend/application-components';
 import history from '@commercetools-frontend/browser-history';
 import type { ApplicationWindow } from '@commercetools-frontend/constants';
 import type { TAsyncLocaleDataProps } from '@commercetools-frontend/i18n';
@@ -8,7 +9,6 @@ import PrimaryButton from '@commercetools-uikit/primary-button';
 import Spacings from '@commercetools-uikit/spacings';
 import Text from '@commercetools-uikit/text';
 import ApplicationShell from '../application-shell';
-import CustomViewLoader from '../custom-view-loader';
 
 declare let window: ApplicationWindow;
